+++ CHANGELOG.md
@@ -4,10 +4,11 @@ Follows the keep a changelog convention and semantic versioning.
 
 ## Unreleased
 ### Added
-- firebase
+- firestore to create and (soft) delete products
 - templating
 - navbar
 - mvpcss styling
 - i18n
 - html component creation with ui lib
 - snackbar
+- sheet for prompting